# server/openapi.yaml
openapi: 3.0.3
info:
  title: Finanzen API
  description: >
    Private Finanzverwaltung mit doppelter Buchhaltung, Kontenrahmen, Zahlungsempfängern,
    wiederkehrenden Zahlungen, Prognosen und CSV-Import.
  version: "1.0.0"
servers:
  - url: /api
...
paths:
...
  /transactions:
    get:
      tags: [Transactions]
      summary: Buchungen auflisten (mit Filtern)
      parameters:
        - name: account
          in: query
          schema: { type: integer }
        - name: payee
          in: query
          schema: { type: integer }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Liste der Buchungen
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Transaction" }
    post:
      tags: [Transactions]
      summary: Neue Buchung erfassen (Postings müssen sich zu 0 summieren)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransactionInput" }
      responses:
        "201":
          description: Buchung erstellt
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IdResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /transactions/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      tags: [Transactions]
      summary: Einzelne Buchung abrufen
      responses:
        "200":
          description: Buchung
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Transactions]
      summary: Buchung ändern
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransactionInput" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Transactions]
      summary: Buchung löschen
      responses:
        "200": { $ref: "#/components/responses/Ok" }
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    Ok:
      description: Erfolgreich
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean }
    NotFound:
      description: Nicht gefunden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    BadRequest:
      description: Ungültige Anfrage
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
    IdResponse:
      type: object
      properties:
        id: { type: integer }
...
    Posting:
      type: object
      properties:
        id: { type: integer }
        accountId: { type: integer }
        accountName: { type: string }
        amountCents:
          type: integer
          description: "Vorzeichenbehaftet: positiv = Geld fließt ins Konto, negativ = Geld fließt heraus"
    PostingInput:
      type: object
      required: [accountId, amountCents]
      properties:
        accountId: { type: integer }
        amountCents: { type: integer }
    Transaction:
      type: object
      properties:
        id: { type: integer }
        date: { type: string, format: date }
        description: { type: string, nullable: true }
        payeeId: { type: integer, nullable: true }
        payeeName: { type: string, nullable: true }
        postings:
          type: array
          items: { $ref: "#/components/schemas/Posting" }
    TransactionInput:
      type: object
      required: [date, postings]
      properties:
        date: { type: string, format: date }
        description: { type: string }
        payeeId: { type: integer, nullable: true }
        payeeName: { type: string, nullable: true }
        postings:
          type: array
          minItems: 2
          items: { $ref: "#/components/schemas/PostingInput" }
          description: "Die Beträge aller Zeilen müssen sich zu 0 summieren."
...